// server/src/api/Types/DatabaseMaps/PartyHistoryDB.ts
export class PartyHistoryDB {
    partyId: string;
    name: string;
    artist: string;
    albumArtwork: string;
    spotifyId: string;
    votes: number;
    addedBy: string;
    addedAt: number;
    playedAt: number;
    acousticness: number;
    danceability: number;
    energy: number;
    instrumentalness: number;
    key: number;
    liveness: number;
    loudness: number;
    mode: number;
    speechiness: number;
    tempo: number;
    time_signature: number;
    valance: number;
}